+++ spec/unit_measurements/measurement_spec.rb
@@ -15,11 +15,11 @@ RSpec.describe UnitMeasurements::Measurement do
 
   describe "#initialize" do
     it "raises an error for blank quantity" do
-      expect { UnitMeasurements::Length.new(nil, "m") }.to raise_error(UnitMeasurements::BaseError, "Quantity cannot be blank.")
+      expect { UnitMeasurements::Length.new("", "m") }.to raise_error(UnitMeasurements::BlankQuantityError, "Quantity cannot be blank.")
     end
 
     it "raises an error for blank unit" do
-      expect { UnitMeasurements::Length.new(1, nil) }.to raise_error(UnitMeasurements::BaseError, "Unit cannot be blank.")
+      expect { UnitMeasurements::Length.new(1, "") }.to raise_error(UnitMeasurements::BlankUnitError, "Unit cannot be blank.")
     end
 
     it "sets attributes correctly" do